Move item interfaces to api module
