+++ code/api/src/main/java/org/betonquest/betonquest/api/item/ItemRegistry.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.item;
+package org.betonquest.betonquest.api.item;
 
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.quest.FeatureTypeRegistry;
