+++ code/api/src/main/java/org/betonquest/betonquest/api/item/QuestItemSerializer.java
@@ -1,7 +1,6 @@
-package org.betonquest.betonquest.item;
+package org.betonquest.betonquest.api.item;
 
 import org.betonquest.betonquest.api.QuestException;
-import org.betonquest.betonquest.api.item.QuestItem;
 import org.betonquest.betonquest.api.quest.TypeFactory;
 import org.bukkit.inventory.ItemStack;
