+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/craftengine/CraftEngineIntegrator.java
@@ -1,11 +1,11 @@
 package org.betonquest.betonquest.compatibility.craftengine;
 
 import org.betonquest.betonquest.api.BetonQuestApi;
+import org.betonquest.betonquest.api.item.ItemRegistry;
 import org.betonquest.betonquest.compatibility.HookException;
 import org.betonquest.betonquest.compatibility.Integrator;
 import org.betonquest.betonquest.compatibility.craftengine.item.CraftEngineItemFactory;
 import org.betonquest.betonquest.compatibility.craftengine.item.CraftEngineQuestItemSerializer;
-import org.betonquest.betonquest.item.ItemRegistry;
 
 /**
  * Integrator for CraftEngine.
