+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/craftengine/item/CraftEngineQuestItemSerializer.java
@@ -2,7 +2,7 @@ package org.betonquest.betonquest.compatibility.craftengine.item;
 
 import net.momirealms.craftengine.bukkit.api.CraftEngineItems;
 import org.betonquest.betonquest.api.QuestException;
-import org.betonquest.betonquest.item.QuestItemSerializer;
+import org.betonquest.betonquest.api.item.QuestItemSerializer;
 import org.bukkit.inventory.ItemStack;
 
 /**
