+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/itemsadder/item/ItemsAdderItemWrapper.java
@@ -5,8 +5,8 @@ import net.kyori.adventure.text.Component;
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.item.QuestItem;
+import org.betonquest.betonquest.api.item.QuestItemWrapper;
 import org.betonquest.betonquest.api.profile.Profile;
-import org.betonquest.betonquest.item.QuestItemWrapper;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
