+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/itemsadder/item/ItemsAdderQuestItemSerializer.java
@@ -2,7 +2,7 @@ package org.betonquest.betonquest.compatibility.itemsadder.item;
 
 import dev.lone.itemsadder.api.CustomStack;
 import org.betonquest.betonquest.api.QuestException;
-import org.betonquest.betonquest.item.QuestItemSerializer;
+import org.betonquest.betonquest.api.item.QuestItemSerializer;
 import org.bukkit.inventory.ItemStack;
 
 /**
