+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/mmogroup/mmoitems/MMOQuestItemFactory.java
@@ -5,9 +5,9 @@ import net.Indyuce.mmoitems.api.Type;
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.instruction.Instruction;
+import org.betonquest.betonquest.api.item.QuestItemWrapper;
 import org.betonquest.betonquest.api.quest.TypeFactory;
 import org.betonquest.betonquest.item.QuestItemTagAdapterWrapper;
-import org.betonquest.betonquest.item.QuestItemWrapper;
 
 /**
  * Factory to create {@link MMOQuestItem}s from {@link Instruction}s.
