+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/nexo/item/NexoItemFactory.java
@@ -2,10 +2,10 @@ package org.betonquest.betonquest.compatibility.nexo.item;
 
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.instruction.Instruction;
+import org.betonquest.betonquest.api.item.QuestItemWrapper;
 import org.betonquest.betonquest.api.quest.TypeFactory;
 import org.betonquest.betonquest.compatibility.nexo.NexoParser;
 import org.betonquest.betonquest.item.QuestItemTagAdapterWrapper;
-import org.betonquest.betonquest.item.QuestItemWrapper;
 
 /**
  * Factory for creating {@link QuestItemWrapper} from Nexo items.
