+++ code/core/src/main/java/org/betonquest/betonquest/api/feature/FeatureRegistries.java
@@ -1,11 +1,11 @@
 package org.betonquest.betonquest.api.feature;
 
+import org.betonquest.betonquest.api.item.ItemRegistry;
 import org.betonquest.betonquest.api.quest.FeatureRegistry;
 import org.betonquest.betonquest.api.quest.npc.NpcRegistry;
 import org.betonquest.betonquest.api.text.TextParserRegistry;
 import org.betonquest.betonquest.conversation.ConversationIOFactory;
 import org.betonquest.betonquest.conversation.interceptor.InterceptorFactory;
-import org.betonquest.betonquest.item.ItemRegistry;
 import org.betonquest.betonquest.notify.NotifyIOFactory;
 import org.betonquest.betonquest.schedule.ActionScheduling;
