+++ code/mc_1_20_6/src/main/java/org/betonquest/betonquest/mc_1_20_6/BundledMC_1_20_6.java
@@ -3,9 +3,9 @@ package org.betonquest.betonquest.mc_1_20_6;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.BetonQuestApi;
 import org.betonquest.betonquest.api.common.component.BookPageWrapper;
+import org.betonquest.betonquest.api.item.ItemRegistry;
 import org.betonquest.betonquest.api.text.TextParser;
 import org.betonquest.betonquest.compatibility.Integrator;
-import org.betonquest.betonquest.item.ItemRegistry;
 import org.betonquest.betonquest.mc_1_20_6.item.UpdatedSimpleItemFactory;
 import org.betonquest.betonquest.mc_1_20_6.item.UpdatedSimpleQuestItemSerializer;
